+++ packages/utils/constants/languages.ts
@@ -1,0 +1,4 @@
+export enum LanguagesEnum {
+  es = 'es',
+  en = 'en',
+}