+++ packages/utils/functions/language/getLanguage/index.ts
@@ -1,0 +1,8 @@
+import { LANGUAGE_COOKIE_NAME } from '../../../constants/cookie'
+import type { ServerSideRenderingOption } from '../../../types/server'
+import { getCookie } from '../../cookie'
+
+export const getLanguage = (
+  cookieName = LANGUAGE_COOKIE_NAME,
+  { noSSR = false }: ServerSideRenderingOption = {},
+) => getCookie<string>(cookieName, { noSSR })